Refactor: self-refactoring(메뉴별 컴포넌트 분리)
